gh_variables.html: Hardcode default branch name to work around bug
The problem is that `gh_variables.html` contains logic to crawl the GitHub org's publicly available repos in order to find the current one and extract information, but isn't robust to errors caused by inaccessible repos (in this case due to a DMCA takedown). Fix by hardcoding the default branch name here instead.